requirements.txt
tqdm
numpy
scipy
torch
deli
transformers
wandb
pyyaml
argparse
biopython
accelerate
timm
rdkit
tensorboard
omegaconf
spyrmsd
prody